add findSFRSymbolAddress, findSFRBitSymbolAddress, isSpecialRegister, some related peephole patterns
the new functions are for looking up the address of an SFR operand and also
detect user defined SFR aliases by standard SFR addresses.  it allows some
peephole patterns to be implemented which would otherwise be unsafe.

see also
SFR bit addressing optimization
https://sourceforge.net/p/sdcc/feature-requests/992/

-> for this, need to have some more checks

currently isSpecialRegister / mcs51isSpecialRegister only checks whether an
operand matches a known standard SFR by name, like acc, dpl, dph, psw.  it will
not match anything that is is not in the compiler's register list.  for this
case, this needs to be separated:

isSpecialRegister: make 2nd parameter optional.  if it's not specified then just
check if the operand is known to be an SFR and don't try to match it to a
particular known compiler-used register.

isBitAddressable: check if an operand is bit addressable.  this could be an SFR
or it could be a variable in the bit-addressable data space (bdata).
if the 2nd optional operand is specified, the bit address will be stored there.
this might be useful in some cases where a immediate bit address is needed.

the peephole would then do:

 if isSpecialRegister(%1), isBitAddressable(%1 %2)